Adding stats helper module
