+++ app/controllers/certification_stats_controller.rb
@@ -1,6 +1,8 @@
 require "json"
 
 class CertificationStatsController < ApplicationController
+  helper StatsHelper
+
   before_action :verify_authentication
   before_action :verify_access
 
@@ -19,28 +21,8 @@ class CertificationStatsController < ApplicationController
     verify_authorized_roles("System Admin")
   end
 
-  def json
-    @stats.map { |d| { key: d.range_start.to_f, value: d.values } }.to_json
-  end
-  helper_method :json
-
   def interval
     @interval ||= CertificationStats::INTERVALS.find { |i| i.to_s == params[:interval] } || :hourly
   end
   helper_method :interval
-
-  # Should be pulled into Caseflow Commons:
-
-  def format_time_duration_stat(seconds)
-    return "?? <span class=\"cf-stat-unit\">sec</span>".html_safe unless seconds
-    return "#{format('%.2f', seconds)} <span class=\"cf-stat-unit\">sec</span>".html_safe if seconds < 60
-    "#{format('%.2f', seconds / 60)} <span class=\"cf-stat-unit\">min</span>".html_safe
-  end
-  helper_method :format_time_duration_stat
-
-  def format_rate_stat(num, denom)
-    "#{@stats[0].values[denom] == 0 ? '??' : (@stats[0].values[num] / @stats[0].values[denom] * 100).round} " \
-      "<span class=\"cf-stat-unit\">%</span>".html_safe
-  end
-  helper_method :format_rate_stat
 end
